requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
requests==2.32.3
openai==1.51.2
celery==5.3.6
redis==5.2.0
pillow==10.4.0
psycopg2-binary==2.9.9